django/requirements.txt
aioredis==1.3.1
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
black==21.11b0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.1.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.0.0
drf-yasg==1.20.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.8.2
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.2
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python-decouple==3.5
pytz==2021.3
regex==2021.11.10
requests==2.26.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
tomli==1.2.2
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typed-ast==1.5.0
typing-extensions==4.0.0
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
zope.interface==5.4.0
